requirements.txt
python-dotenv==1.1.1
aiogram==3.21.0
websockets==15.0.1
aiohttp==3.12.14
pandas==2.3.1